removed static_cast

+++ tasks/seq/sharapov_g_sobel/func_tests/main.cpp
@@ -56,7 +56,7 @@ TEST(sharapov_g_sobel_seq, Test_Seed_1234) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
+    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
   }
 }
 
@@ -110,7 +110,7 @@ TEST(sharapov_g_sobel_seq, Test_Seed_2451) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
+    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
   }
 }
 
@@ -164,7 +164,7 @@ TEST(sharapov_g_sobel_seq, Test_Seed_2423) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
+    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
   }
 }
 
@@ -218,7 +218,7 @@ TEST(sharapov_g_sobel_seq, Test_Seed_2354) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
+    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
   }
 }
 
@@ -272,6 +272,6 @@ TEST(sharapov_g_sobel_seq, Test_Seed_3453) {
   ASSERT_EQ(SSobelSeq.post_processing(), true);
 
   for (size_t i = 0; i < ColoredImage.size(); ++i) {
-    ASSERT_EQ(static_cast<int>(EdgeImageExpected[i].value), static_cast<int>(EdgeImage[i].value));
+    ASSERT_EQ(EdgeImageExpected[i].value, EdgeImage[i].value);
   }
 }
